+++ ci/lib_tezos_ci_jobs/debian_repository.ml
@@ -380,7 +380,6 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
       job_install_systemd_bin
         ~__POS__
         ~name:"oc.install_bin_ubuntu_noble_systemd"
-        ~allow_failure:Yes
         ~dependencies:
           (Dependent
              [
@@ -399,7 +398,6 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
       job_install_systemd_bin
         ~__POS__
         ~name:"oc.upgrade_bin_ubuntu_jammy_systemd_test"
-        ~allow_failure:Yes
         ~dependencies:
           (Dependent
              [
@@ -417,8 +415,25 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
         ];
       job_install_systemd_bin
         ~__POS__
+        ~name:"oc.upgrade_bin_ubuntu_jammy_systemd_test_no_restart"
+        ~dependencies:
+          (Dependent
+             [
+               Job job_docker_systemd_test_ubuntu_dependencies;
+               Job job_apt_repo_ubuntu;
+             ])
+        ~variables:
+          (variables
+             ~kind:"systemd-tests"
+             [("PREFIX", ""); ("DISTRIBUTION", "ubuntu"); ("RELEASE", "jammy")])
+        [
+          "./scripts/ci/systemd-packages-test.sh \
+           scripts/packaging/tests/deb/upgrade-systemd-test-no-restart.sh \
+           images/packages/debian-systemd-tests.Dockerfile";
+        ];
+      job_install_systemd_bin
+        ~__POS__
         ~name:"oc.upgrade_bin_ubuntu_noble_systemd_test"
-        ~allow_failure:Yes
         ~dependencies:
           (Dependent
              [
@@ -432,6 +447,24 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
         [
           "./scripts/ci/systemd-packages-test.sh \
            scripts/packaging/tests/deb/upgrade-systemd-test.sh \
+           images/packages/debian-systemd-tests.Dockerfile";
+        ];
+      job_install_systemd_bin
+        ~__POS__
+        ~name:"oc.upgrade_bin_ubuntu_noble_systemd_test_no_restart"
+        ~dependencies:
+          (Dependent
+             [
+               Job job_docker_systemd_test_ubuntu_dependencies;
+               Job job_apt_repo_ubuntu;
+             ])
+        ~variables:
+          (variables
+             ~kind:"systemd-tests"
+             [("PREFIX", ""); ("DISTRIBUTION", "ubuntu"); ("RELEASE", "noble")])
+        [
+          "./scripts/ci/systemd-packages-test.sh \
+           scripts/packaging/tests/deb/upgrade-systemd-test-no-restart.sh \
            images/packages/debian-systemd-tests.Dockerfile";
         ];
     ]
@@ -454,7 +487,6 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
       job_install_systemd_bin
         ~__POS__
         ~name:"oc.install_bin_debian_bookworm_systemd_test"
-        ~allow_failure:Yes
         ~dependencies:
           (Dependent
              [
@@ -477,7 +509,6 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
       job_install_systemd_bin
         ~__POS__
         ~name:"oc.upgrade_bin_debian_bookworm-systemd"
-        ~allow_failure:Yes
         ~dependencies:
           (Dependent
              [
@@ -495,6 +526,28 @@ let jobs ?(limit_dune_build_jobs = false) pipeline_type =
         [
           "./scripts/ci/systemd-packages-test.sh \
            scripts/packaging/tests/deb/upgrade-systemd-test.sh \
+           images/packages/debian-systemd-tests.Dockerfile";
+        ];
+      job_install_systemd_bin
+        ~__POS__
+        ~name:"oc.upgrade_bin_debian_bookworm-systemd_no_restart"
+        ~dependencies:
+          (Dependent
+             [
+               Job job_docker_systemd_test_debian_dependencies;
+               Job job_apt_repo_debian;
+             ])
+        ~variables:
+          (variables
+             ~kind:"systemd-tests"
+             [
+               ("PREFIX", "");
+               ("DISTRIBUTION", "debian");
+               ("RELEASE", "bookworm");
+             ])
+        [
+          "./scripts/ci/systemd-packages-test.sh \
+           scripts/packaging/tests/deb/upgrade-systemd-test-no-restart.sh \
            images/packages/debian-systemd-tests.Dockerfile";
         ];
     ]